🐛 fix : 멘토/멘티에 따른 회원정보 수정

+++ backend/src/main/java/com/ssafy/foss/member/service/MemberService.java
@@ -98,13 +98,19 @@ public class MemberService {
         Member member = memberRepository.findById(id).orElseThrow(
                 () -> new RuntimeException("해당 식별자를 가진 사용자가 존재하지 않습니다.")
         );
-
-        awsS3Service.deleteFile(member.getProfileImg());
-        String profileImgSrc = awsS3Service.uploadProfile(profileImg);
-        member.change(updateMemberRequest, profileImgSrc);
+        
+        if ("empty-profile-img.png".equals(profileImg.getOriginalFilename())) {
+            member.change(updateMemberRequest);
+        } else {
+            awsS3Service.deleteFile(member.getProfileImg());
+            String profileImgSrc = awsS3Service.uploadProfile(profileImg);
+            member.change(updateMemberRequest, profileImgSrc);
+            member.change(updateMemberRequest);
+        }
 
         return member;
     }
+
 
     @Transactional
     public Member updateRole(Long id) {
